flask_app/requirements.txt
flask
pymongo
dnspython
python-dotenv
pytest
pytest-mock